+++ internal/art/main.go
@@ -47,7 +47,6 @@ func main() {
 		{"statuses", statuses},
 		{"architecture", architecture},
 		{"components", components},
-		{"editor", editor},
 	}
 
 	dir := filepath.Join("docs", "assets")